move decam fixtures
